feat: move util to class
