+++ src/libs/Void.ts
@@ -3,15 +3,16 @@ import type { ConfigObject } from "@open-wa/wa-automate";
 
 import { create } from "@open-wa/wa-automate";
 import MessageHandler from "../handler/Message";
+import Util from "./Util";
 
 export default class Void {
     public constructor(public readonly config: typeof VoidConfig, public readonly options: ConfigObject) {
         void create(options).then(client => {
             const handler = new MessageHandler(client, this.config.prefix);
-            void handler.loadAll();
             client.handler = handler;
             client.config = config;
-            client.wait = async (ms: number): Promise<NodeJS.Timeout> => new Promise(resolve => setTimeout(resolve, ms));
+            client.util = new Util(client);
+            void handler.loadAll();
             void client.onAnyMessage(message => {
                 void client.getAmountOfLoadedMessages().then(msg => msg >= 3000 ? client.cutMsgCache() : msg);
                 void handler.handle(message);
@@ -29,6 +30,6 @@ declare module "@open-wa/wa-automate" {
     interface Client {
         handler: MessageHandler;
         config: typeof VoidConfig;
-        wait(ms: number): Promise<NodeJS.Timeout>;
+        util: Util;
     }
 }
